Keep youtube-dl running preventing process start overhead
Now youtube-dl is started via a small wrapper script that loads
youtube-dl as a daemon, removing the (significant) overhead of starting
a new Python process and initializing the youtube-dl library every time
a video is fetched.
This reduces fetch time from ±14s to ±3s on the Raspberry Pi.
More improvements might be possible, but need to happen in the
youtube-dl library itself. One possible improvement might be keeping
HTTP connections alive and re-using them. This should reduce HTTPS
request time from about 1s to about 40ms for any but the first reqeust.